+++ src/main/resources/mapper/common/delivery/DeliveryMapper.xml
@@ -35,71 +35,77 @@
         <result property="categoryStatus" column="category_status"/>
     </resultMap>
 
+
+
     <!-- 여기부터 작업 -->
     <sql id="deliveryListWhere">
         <where>
-        <!-- 기간 선택 -->
-        <if test="dateSettStartDate != null and dateSettStartDate != ''">
-            DATE_FORMAT(ot.order_application_date, '%Y-%m-%d') >= #{dateSettStartDate}
-        </if>
-        <if test="dateSettEndDate != null and dateSettEndDate != ''">
-            AND
-            #{dateSettEndDate} >= DATE_FORMAT(ot.order_application_date, '%Y-%m-%d')
-        </if>
-        <!-- 공고 분류 선택 -->
-        <if test="projectPartition != null">
-            AND
-            substring(goods_code, 1,3) IN
-            <foreach collection="projectPartition" item="item" open="(" close=")"  separator=",">
-                #{item}
-            </foreach>
-        </if>
-        <!-- 공고 상태 선택 -->
-        <if test="orderby == 'projecting_o'">
-            AND
-            ot.order_category_code IN ('ORDC_001','ORDC_002','ORDC_003')
-        </if>
-        <if test="orderby == 'projectFail_o'">
-            AND
-            ot.order_category_code = 'ORDC_013'
-        </if>
-        <if test="orderby == 'deliveryWait_o'">
-            AND
-            ot.order_category_code = 'ORDC_004'
-        </if>
-        <if test="orderby == 'deliveryCompleted_o'">
-            AND
-            ot.order_category_code = 'ORDC_005'
-        </if>
-        <if test="orderby == 'swapping_o'">
-            AND
-            ot.order_category_code IN ('ORDC_011','ORDC_012')
-        </if>
-        <if test="orderby == 'refunding_o'">
-            AND
-            ot.order_category_code IN ('ORDC_009','ORDC_010')
-        </if>
-        <if test="orderby == 'orderCompleted'">
-            AND
-            ot.order_category_code IN ('ORDC_006','ORDC_007','ORDC_008')
-        </if>
-        <!-- 키워드 선택 -->
-        <if test="searchKey != null and searchKey == 'total'">
-            AND
-            (
-            ot.order_code LIKE CONCAT('%', #{searchValue}, '%')
-            OR
-            ot.member_id LIKE CONCAT('%', #{searchValue}, '%')
-            OR
-            ot.order_name LIKE CONCAT('%', #{searchValue}, '%')
-            OR
-            ot.order_funding_name LIKE CONCAT('%', #{searchValue}, '%')
-            )
-        </if>
-        <if test="searchKey != null and searchKey != 'total' and searchKey != ''">
-            AND
-            ot.${searchKey} LIKE CONCAT('%', #{searchValue}, '%');
-        </if>
+            <!-- 기간 선택 -->
+            <if test="dateSettStartDate != null and dateSettStartDate != ''">
+                DATE_FORMAT(substring(od.order_delivery_code, 6, 8), '%Y-%m-%d') >= #{dateSettStartDate}
+            </if>
+            <if test="dateSettEndDate != null and dateSettEndDate != ''">
+                AND
+                #{dateSettEndDate} >= DATE_FORMAT(substring(od.order_delivery_code, 6, 8), '%Y-%m-%d')
+            </if>
+
+            <!-- 공고 상태 선택 -->
+            <if test="orderby == 'deliveryWait_o'">
+                AND
+                od.delivery_state = 1
+            </if>
+            <if test="orderby == 'delivery_o'">
+                AND
+                od.delivery_state = 2
+            </if>
+            <if test="orderby == 'deliveryCompleted_o'">
+                AND
+                od.delivery_state = 3
+            </if>
+            <if test="orderby == 'swapping_o'">
+                AND
+                od.delivery_category = '교환 배송'
+            </if>
+            <if test="orderby == 'refunding_o'">
+                AND
+                od.delivery_category = '환불 배송'
+            </if>
+
+            <!-- 키워드 선택-->
+            <if test="searchKey != null and searchKey == 'total'">
+                AND
+                (
+                od.order_code LIKE CONCAT('%', #{searchValue}, '%')
+                OR
+                od.member_id LIKE CONCAT('%', #{searchValue}, '%')
+                OR
+                od.delivery_name LIKE CONCAT('%', #{searchValue}, '%')
+                )
+            </if>
+            <if test="searchKey != null and searchKey != 'total' and searchKey != ''">
+                AND
+                od.${searchKey} LIKE CONCAT('%', #{searchValue}, '%');
+            </if>
+        </where>
+
+        <!-- ORDER BY -->
+        <choose>
+            <when test="orderby == 'delived_d'">
+                ORDER BY SUBSTRING(od.order_delivery_code, 6, 8) DESC
+            </when>
+            <when test="orderby == 'delived_a'">
+                ORDER BY SUBSTRING(od.order_delivery_code, 6, 8) ASC
+            </when>
+            <when test="orderby == 'order_name_d'">
+                ORDER BY od.delivery_name DESC
+            </when>
+            <when test="orderby == 'order_name_a'">
+                ORDER BY od.delivery_name ASC
+            </when>
+            <otherwise>
+                ORDER BY SUBSTRING(od.order_delivery_code, 6, 8) DESC
+            </otherwise>
+        </choose>
     </sql>
     <!-- delivery list  조회 -->
     <select id="getDeliveryList" parameterType="map" resultMap="orderDeliveryMap">
@@ -120,9 +126,10 @@
             od.delivery_state,
             od.delivery_start_date,
             od.delivery_end_date,
-            od.delivery_end_date
+            od.due_date_arrival
         FROM
             order_delivery AS od
+        <include refid="deliveryListWhere"></include>
         <if test="startRowNum > -1">
             LIMIT ${startRowNum}, ${rowPerPage};
         </if>
@@ -135,7 +142,7 @@
             COUNT(1)
         FROM
             order_delivery AS od
-
+        <include refid="deliveryListWhere"></include>
     </select>
 
     <!-- 특정 배송 정보 조회 -->
@@ -156,11 +163,19 @@
             od.delivery_courier,
             od.delivery_state,
             od.delivery_start_date,
-            od.delivery_end_date
+            od.delivery_end_date,
+            od.due_date_arrival
         FROM
             order_delivery AS od
-        WHERE
-            od.order_delivery_code = #{orderDeliveryCode};
+        <where>
+            <if test="orderDeliveryCode != null">
+                od.order_delivery_code = #{orderDeliveryCode}
+            </if>
+            <if test="orderCode != null">
+                od.order_code = #{orderCode}
+            </if>
+        </where>
+
     </select>
 
     <select id="getDeliveryCourierCategory" resultMap="deliveryCourierCategoryMap">
