+++ src/main/resources/templates/admin/layout/default.html
@@ -87,7 +87,7 @@
     <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
 
     <!-- Custom Theme Scripts -->
-    <script th:src="@{/admin/build/js/custom.min.js}"></script>
+    <script th:src="@{/admin/build/js/custom.js}"></script>
     <!-- common js -->
     <script th:src="@{/common/js/common-custom.js}"></script>
     <script th:src="@{/admin/js/settlement/settlementJs.js}"></script>
