+++ src/main/resources/templates/common/fragments/deliverInfo.html
@@ -37,8 +37,8 @@
                     배송 완료 예정일
                 </td>
                 <td>
-                    <input type="date" name="dueDateArrival" th:value="${getDeliveryByCode.dueDateArrival}">
-                    <input type="text" style="outline:none; padding: 0; border: none; color:var(--maincolor);" name="dda" readonly="readonly">
+<!--                    <input type="date" name="dueDateArrival" th:value="${getDeliveryByCode.dueDateArrival}">-->
+                    <input type="text" th:value="${getDeliveryByCode.dueDateArrival}" style="outline:none; padding: 5px 0 !important; border: none; color:var(--maincolor);" name="dda" readonly="readonly">
                 </td>
             </tr>
             <tr>
@@ -65,7 +65,7 @@
                     <select name="delivery_category">
                         <option value="" th:selected="${getDeliveryByCode.deliveryCategoryCode == null}">=== 택배사를 선택해주세요 ===</option>
                         <th:block th:each="l : ${deliveryCourierCategory}">
-                            <option th:value="${l.deliveryCategoryCode}" th:selected="${l.deliveryCategoryCode} == ${getDeliveryByCode.deliveryCategoryCode}">[[${l.categoryStatus}]]</option>
+                            <option th:value="${l.deliveryCategoryCode}" th:selected="${l.deliveryCategoryCode} == ${getDeliveryByCode.deliveryCategoryCode}">[[${l.deliveryCourierName}]]</option>
                         </th:block>
                     </select>
                 </td>
@@ -86,7 +86,7 @@
                         <option value="2">배송 전에 연락 바랍니다</option>
                         <option value="3">직접 받고 부재시 문 앞</option>
                         <option value="4">부재시 경비실에 두고 가주세요</option>
-                        <option value="5" >직접 입력</option>
+                        <option value="5">직접 입력</option>
                     </select>
                     <input type="text" width="700px" maxlength="20" placeholder="직접입력" name="deliveryMessageDirect" disabled>
                 </td>
@@ -96,15 +96,15 @@
                     배송 시작일
                 </td>
                 <td style="text-align: left !important;">
-                    <input type="date" name="deliveryStartDate" th:value="${getDeliveryByCode.deliveryStartDate}">
-                    <input type="text" style="outline:none; padding: 0; border: none; color:var(--maincolor);" name="dsd" readonly="readonly">
+<!--                    <input type="date" name="deliveryStartDate" th:value="${getDeliveryByCode.deliveryStartDate}">-->
+                    <input type="text" th:value="${getDeliveryByCode.deliveryStartDate}" style="outline:none; padding: 5px 0 !important; border: none; color:var(--maincolor);" name="dsd" readonly="readonly">
                 </td>
                 <td class="title-td">
                     배송 완료일
                 </td>
                 <td>
-                    <input type="date" name="deliveryEndDate" th:value="${getDeliveryByCode.deliveryEndDate}">
-                    <input type="text" style="outline:none; padding: 0; border: none; color:var(--maincolor);" name="ded" readonly="readonly">
+<!--                    <input type="date" name="deliveryEndDate" th:value="${getDeliveryByCode.deliveryEndDate}">-->
+                    <input th:value="${getDeliveryByCode.deliveryEndDate}" type="text" style="outline:none; padding: 5px 0 !important; ; border: none; color:var(--maincolor);" name="ded" readonly="readonly">
                 </td>
             </tr>
         </table>
