+++ src/main/resources/templates/common/fragments/searchFormFragments.html
@@ -21,7 +21,7 @@
                     <input type="date" name="amDateSettEndDate">
                 </td>
             </tr>
-            <tr>
+            <tr th:if="${paramActive != 'noActive'}">
                 <td>
                     분류
                 </td>
